Allow for longer student IDs.  Our school uses 14 digit bar codes on student cards.

+++ src/main/java/activities/UserActivity.java
@@ -237,7 +237,7 @@ public class UserActivity {
         int mentorIdLength = LocalDbActivity.kIdLengthFallback;
 
         try {
-            Integer.parseInt(userID);
+            Long.parseLong(userID);
 
             if (Constants.kMentorFallback) {
                 return userID.length() == idLength || userID.length() == mentorIdLength;
